pyproject.toml
[project]
name = "learning-tool"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.86.0",
    "fastapi>=0.135.1",
    "google-genai>=1.68.0",
    "jinja2>=3.1.6",
    "numpy>=2.4.3",
    "python-multipart>=0.0.22",
    "pypdf>=5.0.0",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.3.0",
    "typer>=0.12.0",
    "uvicorn[standard]>=0.42.0",
    "alembic>=1.18.4",
    "mcp>=1.26.0",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["learning_tool*"]

[project.scripts]
learn = "learning_tool.cli.main:app"

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [".claude"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["typer.Argument", "typer.Option"]

[tool.ruff.lint.isort]
known-first-party = ["learning_tool"]

[tool.pytest.ini_options]
addopts = "--import-mode=importlib -m 'not slow'"
pythonpath = ["src"]
markers = ["slow: requires real model, run with -m slow"]

[tool.coverage.run]
branch = true
source = ["learning_tool"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
show_missing = true

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.7.0",
    "types-pyyaml>=6.0.12.20250915",
]
